requirements.txt
# python == 3.8.12 #################
# base -----------------------------
loguru
arrow
python-dotenv

# pygame --------------------------
pygame == 2.0.1

# typer ---------------------------
typer